add advanced q

+++ common-base/src/main/java/com/stellariver/milky/common/base/AbstractPageQuery.java
@@ -19,10 +19,10 @@ public abstract class AbstractPageQuery {
 
     @NotNull
     @Positive
-    Long pageIndex;
+    Integer pageIndex;
 
     @NotNull
     @Positive
-    Long pageSize;
+    Integer pageSize;
 
 }
